Nurlana: book detail page
I added book detail page, and linked it to home page, favorite page, and book card. So on home page, when you click on book card it will take you to the book detail page. And when you add any book to favorite on book detail page, it will be visible on both home page and favorite page.